% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLP_attach_packages.R
\name{attach_package}
\alias{attach_package}
\title{Organise temporarily needed packages}
\usage{
attach_package(pkg.cran = NULL, pkg.bioc = NULL,
  source.bioc = "http://bioconductor.org/biocLite.R")
}
\arguments{
\item{pkg.cran}{Character vector with names of CRAN packages to attach}

\item{pkg.bioc}{Character vector with names of Bioconductor packages to attach}

\item{source.bioc}{character with Bioconductor source site (e.g. \code{http://bioconductor.org/biocLite.R})}

\item{pkg}{Character vector with names of packages to be detached}

\item{character.only}{a logical indicating whether name can be assumed to be a character string.}
}
\value{
character vector with package names which have not yet been attached to workspace before call of
        \code{attach_package}. This vector may be used to detach packages not needed any more with \code{detach_package}.
}
\description{
Attach required packages which may be detached after function run.
}
\details{
Help functions to organise package loading. Some functions require huge annotation packages which shall 
not be loaded for the full session. These packages are attached for an individual function and detached 
after function run.
}
\author{
Frank Ruehle
}
