% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEX_diffLimma.R
\name{diffLimma}
\alias{diffLimma}
\title{Differential expression analysis with limma}
\usage{
diffLimma(GEXMTSet, comparisons, p.value.threshold = 0.05,
  adjust.method = "BH", FC.threshold = log2(1.5),
  projectfolder = file.path("GEX/Diff_limma"), projectname = "",
  matchvar = NULL, Symbol.column = "SYMBOL", sampleColumn = "Sample_Name",
  groupColumn = "Sample_Group", useWeights = TRUE, geneAnno2keep = NULL,
  venn_comparisons = NULL, maxHM = 50, HMcexRow = 1.1, HMcexCol = 1.1,
  HMincludeRelevantSamplesOnly = TRUE, color.palette = "heat.colors",
  FC.heatmap.comparisons = NULL, FC.heatmap.geneselection = c("allprobes",
  "intersect", "union"))
}
\arguments{
\item{GEXMTSet}{ExpressionSet or MethylSet}

\item{comparisons}{character vector with group comparisons in format \code{"groupA-groupB"} or 
nested comparisons in format \code{"(groupA-groupB)-(groupC-groupD)"}.}

\item{p.value.threshold}{numeric p-value threshold}

\item{adjust.method}{adjustment method for multiple testing (\code{"none", "BH", "BY" and "holm"})}

\item{FC.threshold}{numeric foldchange threshold. If data is log2-transformed, mind to transform threshold too!}

\item{projectfolder}{character with directory for output files (will be generated if not exisiting).}

\item{projectname}{optional character prefix for output file names.}

\item{matchvar}{(character) if paired design, name of column to be used to group samples. NULL for unpaired design.}

\item{Symbol.column}{character with column name of Gene Symbols in \code{GEXMTSet} or NULL.}

\item{sampleColumn}{character with column name of Sample names in \code{GEXMTSet}}

\item{groupColumn}{character with column name of group names in \code{GEXMTSet}. Group names must match \code{comparisons}!}

\item{useWeights}{(boolean) if TRUE, arrays are weighted for their quality.}

\item{geneAnno2keep}{(character vector) annotation columns in feature data to keep in output}

\item{venn_comparisons}{character vector or (named) list of character vectors with group comparisons to be included 
in Venn Diagram (max 5 comparisons per Venn diagramm)}

\item{maxHM}{(numeric) max number of top diff. regulated elements printed in Heatmap}

\item{HMcexRow}{labelsize for rownames in Heatmap}

\item{HMcexCol}{labelsize for colnames in Heatmap}

\item{HMincludeRelevantSamplesOnly}{(boolean) if TRUE include only Samples in heatmap, 
which are in groups of the respective group comparison.}

\item{color.palette}{select color palette for heatmaps}

\item{FC.heatmap.comparisons}{character vector or (named) list of character vectors with group comparsions to be 
included in foldchange heatmaps. One FC heatmap is generated for each character vector.
If NULL no FC heatmaps are generated.}

\item{FC.heatmap.geneselection}{character vector containing elements of \code{c("all", "intersect", "union")}. 
Which probes shall be used for foldchange heatmaps? \code{"allprobes"}: no restriction,  
\code{"intersect"}: only probes are used which are differentially expressed in all respective group comparisons.
\code{"union"}: only probes are used which are differentially expressed in any of the respective group comparisons.
Probes are prioritised by F-Test p-value calculated for all group comparisons denoted in \code{comparisons}
as well as by minimum adjusted p-value of respective group comparisons plotted in this heatmap.
Obsolete if 'FC.heatmap.comparisons' is NULL.}
}
\value{
List of dataframes with unfiltered differential expression data for each comparison (sorted by p-value).
List elements are named by the respective group comparison. 
An F-Test and a Venn Diagramm of all comparisons as well as filtered and unfiltered result tables and heatmaps 
for each group comparison are stored as side-effects in the project folder.
}
\description{
Differential Gene expression analysis for multiple group comparisons incl. output of Venn diagrams and heatmaps.
}
\details{
Function generates design matrix, fit and differential expression results for dedicated group comparisons of the input object.
Analysis is performed either for one group comparison in paired design or for any number of unpaired group comparisons.
The group designations in the \code{groupColumn} must match the contrasts given in \code{comparisons} (\code{"groupA-groupB"}).
P-value and foldchange thresholds may be applied to filter result data. 
Heatmaps with sample signal intensities are generated for top differentially expressed genes for each group comparison 
given in \code{comparisons}. Additionally heatmaps indicating foldchanges are generated for each set of group comparisons
given in \code{FC.heatmap.comparisons}. Selection of probes with respect to these group comparisons is characterised in 
\code{FC.heatmap.geneselection}. If the resulting number of probes exceeds \code{maxHM}, probes are prioritized by either 
F-Test p-value or by minimum p-value of selected group comparisons.
}
\author{
Frank Ruehle
}
