% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GT_genoAssoc.R
\name{genoAssoc}
\alias{genoAssoc}
\title{Association analysis with GenABEL}
\usage{
genoAssoc(gwaa, projectfolder = "GT/Assoc_GenAbel", projectname = "",
  trait.name = "affection01 ~ Age + sex", trait.type = "binomial",
  times = 1, quiet = FALSE, TopHitsReported = 1000)
}
\arguments{
\item{gwaa}{gwaa object from GenABEL}

\item{projectfolder}{character containing path to output folder (will be generated if not existing).}

\item{projectname}{character used as suffix for output files.}

\item{trait.name}{character indicating trait(s) of interest.}

\item{trait.type}{character with data type of analysed trait. Either "gaussian", "binomial" or "guess" 
(later option guesses trait type).}

\item{times}{If more than one, the number of permutations to be used for empirical p-values.}

\item{quiet}{boolean. Do not print warning messages.}

\item{TopHitsReported}{numeric with number of "top" hits to describe. Use nsnps(gwaa.object) for
all SNPs}
}
\value{
Association results as scan.gwaa object. Intermediary results and plots are stored in \code{projectfolder} 
as side effects.
}
\description{
\code{genoAssoc} takes genotype data in GenABEL gwaa format and performs association analysis.
}
\details{
The qtscore()-function from GenABEL package is used for association analysis of a gwaa object.
The trait of interest is given in \code{trait.name} either as single trait (e.g. "affection")
or as formula (e.g. "affection01 ~ Age + sex") if association has to be adjusted for covariates.
All traitnames in \code{trait.name} must exist within the phenotype data of the gwaa object. 
Use phdata() for checking which trait names ara avalable. If \code{times} is set to more than 1,
empirical p-values are calculated using \code{times} permutations.
A table and a Manhattan plot with association results are stored in \code{projectfolder}.
}

