% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GT_genoQC.R
\name{genoQC}
\alias{genoQC}
\title{Quality Control for genotype data}
\usage{
genoQC(gwaa, projectfolder = "GT/QC", projectname = "QC1",
  trait.name = "affection01", trait.type = "binomial",
  export.genofile = "ped", p.level.hwe = 0.05, hwe.id.subset = T,
  maf = 0.01, checkX = T, PCA = T, maxCenters = 5, ...)
}
\arguments{
\item{gwaa}{gwaa object from GenABEL}

\item{projectfolder}{character containing path to output folder (will be generated if not existing).}

\item{projectname}{character used as suffix for output files.}

\item{trait.name}{character indicating column name with trait of interest in pheno data of \code{gwaa}. 
Needed for qq-plots as well as for PCA plot annotation. Omitted if NULL.}

\item{trait.type}{character with data type "gaussian" or "binomial" of \code{trait.name}.}

\item{export.genofile}{character or character vector with type(s) of QC-purified ped file to export into \code{projectfolder}.
Allowed values are "ped" for ped/map-file, "tped" for transposed ped file format or "add.tped" for 
transposed additive coded format. If NULL, no data is exported.}

\item{p.level.hwe}{numeric cut-off p-value for HWE in check.markers. For first round of QC it is rcommended to
skip p-level cut-off, i.e. set \code{p.level.hwe} = 0.}

\item{hwe.id.subset}{Subset for HWE checks in check.markers (default means controls only 
if \code{trait.name} is 0/1-coded affection status).}

\item{maf}{numeric cut off for minor allele frequency to be used in check.markers.}

\item{checkX}{boolean. If TRUE, X-errors in \code{gwaa} are fixed by Xfix().}

\item{PCA}{boolean. IF TRUE, PCA analysis performed with genotype data.}

\item{maxCenters}{numeric with maximum count of reported clustering center if PCA is performed.}

\item{...}{further parameter submitted to GenABEL's check.marker() function. See \code{?check.marker} for details.}
}
\value{
list containing two objects. First the QC-purified GenABEL gwaa object whith all samples  
removed as recomended. Second an object of class check.marker containing the quality control information.
Intermediary results and plots are stored in \code{projectfolder} as side effects.
}
\description{
\code{genoQC} takes genotype data in GenABEL gwaa format and performs quality control and PCA analysis.
}
\details{
The check.marker-function from GenABEL package is used for quality control of genotype data.
It is recommended to perform two round of quality control: first QC, remove samples with different genetic
substructure, second QC. Principal component analysis for detection of genetic substructure is done
if \code{PCA} = TRUE. The first 10 principal components are added to the covariates of the \code{gwaa} object.
Samples are assigned to clusters and colored accordingly in PCA plots. Sample assignment is done for
up to \code{maxCenters} cluster centers. All cluster sample lists are stored in a subfolder "ClusterLists".
The QC-purified gwaa object may be exported to PLINK-compatible file formats.
}

