% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gex}
\alias{gex}
\title{expression data}
\format{An ExpressionSetIllumina object containing 12 samples and 48185 features.
Expression data is artificial for demonstration purposes only and does not 
reflect any biological insights. Phenotype data include \code{Sample_Group} for sample grouping,
\code{age}, \code{gender} (1= male, 2=female) as well as \code{Sentrix_ID} and \code{Sentrix_Position} 
giving the microarray allocation.}
\usage{
gex
}
\description{
Artificial expression data for 12 probands for demonstration purposes in \code{systemsbio}
}
\keyword{datasets}
