% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SYB_matrixQTL.R
\name{matrixQTL}
\alias{matrixQTL}
\title{QTL Analysis for expression and/or methylation data.}
\usage{
matrixQTL(eset = NULL, mset = NULL, islandset = NULL,
  SNPfile_transpAddCoded_name, SNPfile_tfam = NULL,
  covariates_file_name = NULL, covarsampleID = "IID", covar2adjust = NULL,
  projectfolder = "QTL", projectname = NULL, sampleGEX = "Sample_Name",
  sampleMT = "Sample_Name", useModel = modelLINEAR,
  errorCovariance = numeric(), QTLtype = "both", cisDist = 1e+06,
  pvOutputThreshold_cis = 1e-06, pvOutputThreshold_tra = 1e-06,
  pvOutputThreshold_all = 1e-06, genepos = NULL,
  genemart = useMart("ENSEMBL_MART_ENSEMBL", host =
  "feb2014.archive.ensembl.org", dataset = "hsapiens_gene_ensembl"),
  ensembl_filter = "illumina_humanht_12_v4", updateSNPpos = FALSE,
  snpmart = useMart("ENSEMBL_MART_SNP", host = "feb2014.archive.ensembl.org",
  dataset = "hsapiens_snp"))
}
\arguments{
\item{eset}{ExpressionSet object.}

\item{mset}{MethylationSet object.}

\item{islandset}{data from summarized methylation Island (output from COHCAP.avg.by.island).}

\item{SNPfile_transpAddCoded_name}{character with path to genotype file (transposed and additive coded!).}

\item{SNPfile_tfam}{character with path to tfam file with sample information. 
If NULL, sample names are expected as header line of \code{SNPfile_transpAddCoded_name}. 
If given, \code{SNPfile_transpAddCoded_name} is read as no-header file and sample names are added 
from \code{SNPfile_tfam} as column names. Check in advance if you need \code{SNPfile_tfam} or not.}

\item{covariates_file_name}{character with path to covariates file.}

\item{covarsampleID}{Character with column name of sample IDs in covar file.}

\item{covar2adjust}{Character vector with column names of covariates to adjust QTL analysis. Omitted if NULL.}

\item{projectfolder}{Character containing path to output folder (will be generated if not existing).}

\item{projectname}{Character used as suffix for output files.}

\item{sampleGEX}{Character with column name of sample IDs in phenotype data of \code{eset}.}

\item{sampleMT}{Character with column name of sample IDs in phenotype data of \code{mset} or \code{islandset}.}

\item{useModel}{model to use (modelANOVA or modelLINEAR or modelLINEAR_CROSS). 
Set useModel = modelLINEAR to model the effect of the genotype as additive linear and test for its 
significance using t-statistic. Set useModel = modelANOVA to treat genotype as a categorical variables 
and use ANOVA model and test for its significance using F-test. The default number of ANOVA categories is 3.
Set otherwise like this: options(MatrixEQTL.ANOVA.categories=4). Set useModel = modelLINEAR_CROSS to add 
a new term to the model equal to the product of genotype and the last covariate; the significance of this 
term is then tested using t-statistic.}

\item{errorCovariance}{Numeric error covariance matrix. Use numeric() for homoskedastic independent errors.}

\item{QTLtype}{Character with "both" for calculating cis and trans QTL separately, 
"all" for no separation of cis and trans, "cis" for cis QTLs only.}

\item{cisDist}{Numeric maximal baisepair distance for cis gene-SNP pairs.}

\item{pvOutputThreshold_cis}{Numeric significance threshold p-value for cis QTL tests.}

\item{pvOutputThreshold_tra}{Numeric significance threshold p-value for trans QTL tests.}

\item{pvOutputThreshold_all}{Numeric significance threshold p-value for all QTL tests (cis and trans mixed).}

\item{genepos}{Character with path to gene position file for expression data. If NULL, genepos file is 
generated with biomaRt from eset probe IDs.}

\item{genemart}{biomaRt object to be used for updating gene positions.}

\item{ensembl_filter}{Character with filter name to search in genemart.}

\item{updateSNPpos}{Boolean. Shall SNP-positions be updated via biomaRt?}

\item{snpmart}{biomaRt object to be used for updating SNP positions.}
}
\value{
list containing analysis parameter and QTL results. Intermediary results and plots 
are stored in \code{projectfolder} as side effects.
}
\description{
\code{matrixQTL} uses MatrixEQTL package for QTL analysis of an ExpressionSet or/and a MethylationSet object with SNP genotypes.
}
\details{
Expression and methylation data can be processed simulatanously within a single function call. Coordinates 
for genes and SNPs are either given in \code{SNPfile_transpAddCoded_name} or \code{genepos}, respectively,
or are downloaded from biomaRt. Model type and optional covariates can be selected for analysis. 
Analysis can be perfomed for either cis or trans QTLs or both simulatanously.
}
\author{
Frank Ruehle
}
