% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLP_plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Generating Volcano plot for differential gene expression results}
\usage{
plot_volcano(data, column_name_geme = NULL, column_name_p = "padj",
  column_name_fc = "log2FoldChange", title = "Volcano plot",
  p_value_threshold = 0.05, fc_threshold = log2(1.5),
  xlabel = "log fold change", ylabel = "-log10 p-value (adjusted)",
  color_down_reg = "red", color_up_reg = "darkgreen",
  color_not_reg = "darkgray")
}
\arguments{
\item{data}{data.frame containing gene ID or Symbol (optional), p-value and logarithmized foldchange.
Column names can be given in \code{column_name_geme}, \code{column_name_p} and \code{column_name_fc}, respectively.}

\item{column_name_geme}{character with optional column name of gene names or symbols in \code{data}.}

\item{column_name_p}{character with column name of p-values in \code{data}.}

\item{column_name_fc}{character with column names of logarithmized fold changes in \code{data}.}

\item{title}{character with title of volcano plot.}

\item{p_value_threshold}{numeric significance threshold for p-value.}

\item{fc_threshold}{numeric significance threshold for logarithmized fold changes.}

\item{xlabel}{character with label for x-axis}

\item{ylabel}{character with label for y-axis}

\item{color_down_reg}{character with color name for indicating down regulated genes.}

\item{color_up_reg}{character with color name for indicating up regulated genes.}

\item{color_not_reg}{character with color name for indicating genes not differentially regulated.}
}
\value{
dataframe with filtered gene list. Volcano plot is generated as side effect.
}
\description{
Function uses result gene list from differential expression analysis to 
prepare volcano plot.
}
\details{
Input gene list is filtered for differentially expressed genes according to given significance thresholds
for p-value and log foldchange. Significance thresholds are indicated as dashed lines in the plot.
Up and down regulated genes can be colored accordingly. The counts of significantly differentially expressed
genes is given in the upper plot margin.
}
\author{
Frank Ruehle
}
