% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEX_process_dds.R
\name{process_dds}
\alias{process_dds}
\title{Transform DESeqDataSet to SummarizedExperiment}
\usage{
process_dds(dds, min_rowsum = 10, transform.function = NULL,
  norm_sizeFactors = FALSE, export = "SummarizedExperiment", ...)
}
\arguments{
\item{dds}{\code{DESeqDataSet}}

\item{min_rowsum}{numeric. Minimum rowsum of countmatrix. All rows with \code{rowSums < min_rowsum}
are removed from the count matrix.}

\item{transform.function}{function name to be used for data transformation, 
e.g. \code{rlog}, \code{rlogTransformation}, \code{varianceStabilizingTransformation}.}

\item{norm_sizeFactors}{logical. If TRUE normalize for library size. This may already be included
in the data transformation (e.g. for \code{rlog} and \code{varianceStabilizingTransformation}).}

\item{export}{return either a \code{"SummarizedExperiment"} or a \code{"DESeqDataSet"} object.}

\item{...}{further parameter for \code{transform.function}}
}
\value{
\code{SummarizedExperiment} or \code{DESeqDataSet} object
}
\description{
Process count matrix of DESeqDataSet to SummarizedExperiment to be used as expression matrix.
}
\details{
The count matrix of the input object is filtered for low abundancy transcripts, normalized for 
library size factor (size factors are  first calculated if necessary) and is log2-transformed
using the \code{DESeq2::rlog} function. The design parameter of the initial \code{DESeqDataSet}
is omitted
}
\author{
Frank Ruehle
}
