% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEX_process_eset.R
\name{process_eset}
\alias{process_eset}
\title{Processing of expression data}
\usage{
process_eset(eset, method_norm = "quantile", transform = "none")
}
\arguments{
\item{eset}{ExpressionSet or ExpressionSetIllumina}

\item{method_norm}{character with normalisation method. Options are "quantile", "qspline", "vsn", "rankInvariant", "median" and "none"}

\item{transform}{character with data transformation method. Options are "none", "log2", "neqc", "rsn" and "vst".}
}
\value{
processed ExpressionSet object
}
\description{
Normalisation, transformation and/or probe quality filtering of expression data.
}
\details{
Expression data is normalised and/or transformed by algorithms dedicated in 'method_norm' and/or 'transform'.
If a column \code{PROBEQUALITY} is available within the feature data as for Illumina expression arrays, 
probes assigned a `Bad' or `No match' quality score after normalisiation are removed from the dataset.
Lastly, if the input object is of class \code{ExpressionSetIllumina}, it is changed to \code{ExpressionSet}
to avoid later incompabilities with the 'limma'-package.
}
\author{
Frank Ruehle
}
