% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MT_process_meth.R
\name{process_meth}
\alias{process_meth}
\title{Preprocessing and annotation of methylation array data}
\usage{
process_meth(RGset, projectfolder = "MT", projectname = NULL,
  preprocessing = c("illumina", "swan"), addAnno = T,
  SNPloci2drop = c("CpG", "SBE"), MAF2drop = 0,
  return.EpiSNPs.only = F)
}
\arguments{
\item{RGset}{RGChannelSet object with methylation data.}

\item{projectfolder}{character with directory for output files (will be generated if not exisiting).}

\item{projectname}{optional character prefix for output file names.}

\item{preprocessing}{character with preprocessing procedures to apply. Any of \code{"raw", "illumina", "swan"}.
If multiple procedures are selected, the one with higest priority is used: \code{"raw" < "illumina" < "swan"}.
If "swan" is selected, at least one more procedure must be selected, too.
If \code{NULL}, input object is expected to be already preprocessed.}

\item{addAnno}{logical. If TRUE, annotation data is added from annotation package.}

\item{SNPloci2drop}{character with SNP types, either \code{"CpG"} (SNPs overlapping the CpG site) and/or \code{"SBE"} 
(SNP at the single base extension of CpG-site). CpG sites overlapping these SNPs are removed from the dataset. 
If \code{NULL}, no CpG sites are removed.}

\item{MAF2drop}{numeric with minimum minor allele frequency of SNPs to drop if \code{SNPloci2drop} is specified.}

\item{return.EpiSNPs.only}{logical. If TRUE, the returned GenomicRatioSet contains only CpG-sites overlapping with SNP locations
(usefull to focus on SNPs influencing methylation processes).}
}
\value{
GenomicRatioSet (containing Beta and M-values)
}
\description{
An RGChannelSet object is preprocessed, mapped to the genome and annotated by the respective annotation package.
}
\details{
This function takes an RGChannelSet as input and transforms it to a MethylSet object using one of 3 possible preprocessing procedures 
from \code{minfi} package:
\itemize{
  \item \code{preprocessRaw} (no normalisiation)
  \item \code{preprocessIllumina} applies background subtraction as well as control normalization like in Illumina Genome studio
  \item \code{preprocessSWAN} applies SWAN normalization (Subset quantile Within-Array Normalization) method on a preprocessed 
                              MethylSet to reduce the technical variability of Type I and Type II Illumina probes. 
                              For this, Illumina preprocessed MethylSet is used if given, the raw preprocessed MethylSet otherwise.  
  }
If no preprocessing is selected, the input object is expected to be already preprocessed. Next, the MethylSet object is
mapped to the genome (Loci which cannot be mapped to a genomic position are dropped) and converted to a GenomicRatioSet. 
Finally, annotation data is added from the respective annotation package.
CpG sites overlapping with SNPs optionally can be removed from the dataset. If these sites are of special interest, 
the returned \code{GenomicRatioSet} can be specified to contains these CpG-sites only.
}
\author{
Frank Ruehle
}
