% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLP_LNCipedia.R
\name{processLNCipedia}
\alias{processLNCipedia}
\title{Process LNCipedia bed-file}
\usage{
processLNCipedia(LNCipedia, collapseTranscripts2Genes = T,
  makeExonRanges = T, addBases = 0)
}
\arguments{
\item{LNCipedia}{Either character with filepath to LNCipedia bed-file (no header), dataframe 
or GRanges object made from LNcipedia bed-file.}

\item{collapseTranscripts2Genes}{logical. If TRUE, all transcripts of a gene are assigned with the coordinates
of the corresponding gene. Recommended for plotting to reduce complexity.}

\item{makeExonRanges}{logical. If True, an GRanges object for exons is generated using the 
"blockStarts" and "blockSizes" data from LNCipedia. Transcript or gene coordinates are stored in metadata.}

\item{addBases}{numeric. Add bases to both sides of the ranges.}
}
\value{
GRanges object containing LNCipedia transcripts or exons
}
\description{
Load and/or modify lncRNA data from LNCipedia.
}
\details{
LNCipedia date is loaded from bed file or provided as dataframe or GRanges object.
For reducing complexity in plotting purposes, all transcripts belonging to a lncRNA gene are given 
with respective gene coordinates. For each transcript, exon ccordinates are calculated from
"blockStarts" and "blockSizes" data from LNCipedia.
}
\author{
Frank Ruehle
}
