% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEX_QC_expressionset.R
\name{QC_expressionset}
\alias{QC_expressionset}
\title{Quality Control procedures for expression data}
\usage{
QC_expressionset(eset, projectfolder = "GEX/QC", projectname = "",
  groupColumn = "Sample_Group", phDendro = "Sample_Group",
  flashClustMethod = "average", cex.dendroLabels = 0.6, thresholdZ.k = -5)
}
\arguments{
\item{eset}{ExpressionSet or ExpressionSetIllumina object}

\item{projectfolder}{character vector with name of desired output folder (will be generated if not exisiting).}

\item{projectname}{character vector containing prefix for names of output files.}

\item{groupColumn}{character vector with name of group column in phenotype data.}

\item{phDendro}{character vector with phenotypes of eset object to be displayed in sample Dendrogram}

\item{flashClustMethod}{character vector with the agglomeration method to be used in \code{WGCNA} package. 
Can be one of "ward", "single", "complete", "average", "mcquitty", "median" or "centroid".}

\item{cex.dendroLabels}{numeric. Controls size of labels sample in dendrogram.}

\item{thresholdZ.k}{numeric. Threshold for outlier prediction WGCNA package (outlier samples are highlighted 
in red in sample dendrogram).}
}
\value{
The returned value is a list of 3
- goodGenes: boolean vector indicating genes passing QC-criteria as implemented in \code{WGCNA}.
- goodSamples: boolean vector indicating samples passing QC-criteria as implemented in \code{WGCNA}.
- allOK: boolean. If TRUE all samples and all genes pass QC-criteria as implemented in \code{WGCNA}.
Several plots and files are generated as side-effects and stored in the output folder.
}
\description{
Application of diverse quality control metrix for gene expression data.
}
\details{
Function implements several approaches for quality control of expression data, which
are stored in a newly generated 'project folder.
- Illumina internal control plots, box-plots and MA-plots (if applicable).
- Quality control as implemented in the \code{arrayQualityMetrics}-package (subfolder 'quality_metrics')
- Quality control routines implemented in \code{WGCNA}-package. This comprises a sample dendrogram
  with included outlier prediction as well as boolean vectors for sample and gene quality
  calculated by the \code{goodSamplesGenes}-function.
}
\author{
Frank Ruehle
}
