% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MT_QC_methylation.R
\name{QC_methylation}
\alias{QC_methylation}
\title{Quality control of methylation array data}
\usage{
QC_methylation(RGset, projectfolder = "MT/QC", projectname = NULL,
  sampleColumn = "Sample_Name", groupColumn = "Sample_Group",
  preprocessing = c("raw", "illumina", "swan"), QC_procedures = c("minfi",
  "WGCNA"), numPositionsMDSplot = 1000, methPlatform = "450k-UCSC",
  phDendro = "Sample_Group", flashClustMethod = "average",
  cex.dendroLabels = 0.6, threshold.Z.k = -5)
}
\arguments{
\item{RGset}{RGChannelSet object with methylation data.}

\item{projectfolder}{character with directory for output files (will be generated if not exisiting).}

\item{projectname}{optional character prefix for output file names.}

\item{sampleColumn}{character with column name of Sample names in \code{RGset}.}

\item{groupColumn}{character with column name of group names in \code{RGset}.}

\item{preprocessing}{character with preprocessing procedures to apply. Any of \code{"raw", "illumina", "swan"}.
If "swan" is selected, at least one more procedure must be selected to.}

\item{QC_procedures}{character with desired QC procedures to use. Any of \code{"minfi", "COHCAP", "WGCNA"}.}

\item{numPositionsMDSplot}{number of genomic positions with the most methylation variability to be used for 
calculating distance between samples in MDS plots (needed for \code{minfi} only).}

\item{methPlatform}{Annotation file to be used. Enter "450k-UCSC" for UCSC CpG Islands for 450k array probes, "
450k-HMM" for HMM CpG Islands for 450k array probes, "27k" for UCSC CpG Islands for 27k array probes
(needed for \code{COHCAP} only).}

\item{phDendro}{character vector with phenotypes of \code{RGset} to be displayed in sample dendrogram 
(needed for \code{WGCNA} only).}

\item{flashClustMethod}{character vector with the agglomeration method to be used in \code{WGCNA} package. 
Can be one of "ward", "single", "complete", "average", "mcquitty", "median" or "centroid"
(needed for \code{WGCNA} only).}

\item{cex.dendroLabels}{numeric. Controls size of labels sample in dendrogram (needed for \code{WGCNA} only).}

\item{threshold.Z.k}{numeric. Threshold for outlier prediction in \code{WGCNA} package. Outlier samples are highlighted 
in red in sample dendrogram (needed for \code{WGCNA} only).}
}
\value{
no value returned. QC plots are stored as side-effects.
}
\description{
Multiple QC procedures of an RGChannelSet object containing methylation data.
}
\details{
This function takes an RGChannelSet as input and transforms it to a MethylSet object using 3 preprocessing procedures 
from \code{minfi} package:
\itemize{
  \item \code{preprocessRaw} (no normalisiation)
  \item \code{preprocessIllumina} applies background subtraction as well as control normalization like in Illumina Genome studio
  \item \code{preprocessSWAN} applies SWAN normalization method on a preprocessed MethylSet to reduce the technical variability
                              of Type I and Type II Illumina probes. For this, Illumina preprocessed MethylSet is used if given,
                              the raw preprocessed MethylSet otherwise.  
  }
Quality control functions are applied to all 3 MethylSet objects. 
The required annotation packages (currently hg19 only) \code{IlluminaHumanMethylation450kmanifest} and 
\code{IlluminaHumanMethylation450kanno.ilmn12.hg19} are installed and/or loaded automatically. 
Required packages not already attached when running \code{QC_methylation} are detached afterwards.
The function generates the following plots using quality control functionalities from up to 3 packages:
\itemize{
  \item \code{minfi}
    \itemize{
      \item qcReport using array control probes (for RGChannelSet)
      \item Median intensity plots
      \item Predicted sex plots
      \item Multi-dimensional scaling (MDS) plots
      \item Histogram overview of Illumina and SWAN normalisation
     }
  \item \code{COHCAP} (needs much computational resources)
   \itemize{
    \item Cluster plot
    \item PCA plot
    \item histogram plot
     }
  \item \code{WGCNA}
   \itemize{
    \item clustered sample dendrogram
    \item indicated outlier detection
     }
   }
}
\author{
Frank Ruehle
}
