% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MT_read_Illu_meth450k.R
\name{read_Illu_meth450k}
\alias{read_Illu_meth450k}
\title{Reading methylation data from Illumina array HumanMethylation450k}
\usage{
read_Illu_meth450k(baseDir, SampleSheetRegExp = ".csv",
  sampleColumn = "Sample_Name", covarfile = NULL, covarsampleID = "ID")
}
\arguments{
\item{baseDir}{character with path to base directory containing IDAT files for all arrays.}

\item{SampleSheetRegExp}{dataframe with sample sheet or character with regular expression for identifying 
the corresponding sample sheet within the base directory or its subdirectories. If NULL, no sample information file is loaded.}

\item{sampleColumn}{Name of sample column in sample sheet}

\item{covarfile}{character with filepath to a covariates file if given. NULL otherwise.}

\item{covarsampleID}{character with name of sample column in (optional) covariates file.}
}
\value{
RGChannelSet object
}
\description{
Create an RGChannelSet object from Illumina IDAT files and add phenotype data.
}
\details{
This functions uses the \code{minfi} package to load methylation data from Illumnia 
Human Methylation 450k microarrays. 
If the \code{SampleSheetRegExp} argument is NULL, the function finds all two-color IDAT files in the 
directory given by \code{baseDir} including all subdirectories. Two-color IDAT files consist of pairs of 
files with names ending in \code{_Red.idat} or \code{_Grn.idat}. If \code{SampleSheetRegExp} is not NULL,
the corresponding sample sheet is identified via regular expression (file must be located in base) 
and all listed arrays are loaded. 
Susequently, additional phenotype data given in \code{covarfile} is added to the object.
}
\note{
methylation data classes (from \code{minfi} manual) :
\itemize{
  \item RGChannelSet: raw data from the IDAT files; this data is organized at the probe (not CpG locus) level. 
  This data has two channels: Red and Green. This data type contains control probe data.
  \item MethylSet: data organized by the CpG locus level, but not mapped to a genome. 
  This data has two channels: Meth (methylated) and Unmeth (unmethylated).
  \item RatioSet: data organized by the CpG locus level, but not mapped to a genome. The data has at least 
  one of two channels: Beta and/or M (logratio of Beta). It may optionally include a CN channel (copy number).
  \item GenomicMethylSet: like a MethylSet, but mapped to a genome.
  \item GenomicRatioSet: like a RatioSet, but mapped to the genome.
  }
}
\author{
Frank Ruehle
}
