% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GT_readGeno.R
\name{readGeno}
\alias{readGeno}
\title{Read Genotype Data Using GenABEL.}
\usage{
readGeno(genoFilename, mapFilename, covarFilename, existingRawFile = NULL,
  covarsampleID = "IID", sexvar = "Gender", projectfolder = "GT",
  projectname = NULL, organism = "human", Gonosomes2char = T,
  removeNullPositions = T, updateSNPpos = NULL,
  snpmart = useMart("ENSEMBL_MART_SNP", host = "feb2014.archive.ensembl.org",
  dataset = "hsapiens_snp"), GTarrayDescriptionFile = NULL,
  GTarrayDescription.rows2skip = 7,
  GTarrayDescription.colname.identifier = "Name",
  GTarrayDescription.colname.coding = "SNP",
  GTarrayDescription.colname.strand = "RefStrand",
  GTarrayDescription.colname.chromosome = "Chr",
  GTarrayDescription.colname.position = "MapInfo")
}
\arguments{
\item{genoFilename}{Character with path to ped file.}

\item{mapFilename}{Character with path to map file.}

\item{covarFilename}{Character with path to covariates file (gender information is needed).}

\item{existingRawFile}{Character with path to an optionally already existing GenABEL Raw which will preferably loaded
and supersedes any other files. Omitted if NULL.}

\item{covarsampleID}{Character with column name of sample IDs in covar file.}

\item{sexvar}{Character with column name indicating gender in covar file.}

\item{projectfolder}{Character containing path to output folder (will be generated if not existing).}

\item{projectname}{Character used as suffix for output files.}

\item{organism}{Character with name of organism (e.g. "human").}

\item{Gonosomes2char}{Boolean. If TRUE, chromosome names 23, 24, 25, 26 will be converted to X, Y, XY, MT (for human only).}

\item{removeNullPositions}{Boolean. If TRUE, SNPs with Chr=0 or Pos=0 or NA will be removed from generated \code{gwaa} object}

\item{updateSNPpos}{Character with value "biomaRt" or "descriptionfile". For "biomaRt", chromosome and SNP positions 
in map-file will be updated by biomaRt. If "descriptionfile", they will be updated by the supplied \code{GTarrayDescriptionFile}.
If NULL chromosome and SNP positions are not updated.}

\item{snpmart}{biomaRt object to be used for updating SNP positions.}

\item{GTarrayDescriptionFile}{Optional character with path to Illumina array description file. If given, 
strand and allele coding data will be added to map file.}

\item{GTarrayDescription.rows2skip}{Numeric indicating description rows in Array description file to be skipped.
Omitted if \code{GTarrayDescriptionFile} is NULL.}

\item{GTarrayDescription.colname.identifier}{Character with colnames for SNP identifier in Array description file.}

\item{GTarrayDescription.colname.coding}{Character with colnames for allele coding information}

\item{GTarrayDescription.colname.strand}{Character with colnames for strand information.}
}
\value{
GenABEL gwaa object
}
\description{
\code{readGeno} reads genotype data stored in ped/map files, adjusts format for GenABEL package and processes it for further analysis.
}
\details{
If a GenABEL raw data file given in \code{existingRawFile} already exists, this file is directly loaded into a gwaa 
object. Otherwise this file has to be created first. For this, the map file in \code{mapFilename} is loaded, the 
linkage column (column 3) removed if present and a headerline is added. If an Array description file is supplied 
in \code{GTarrayDescriptionFile}, an extended map file is generated for GenABEL including strand and allele coding 
information. Otherwise strand information is set to unknown. If a biomaRt object is given in \code{snpmart} SNP 
chromosomal postions can be updated according either to biomaRt or to \code{GTarrayDescriptionFile}.

After loading the GenABEL raw file, phenotype information is added from the supplied covariates file \code{covarFilename}.
I.e. no phenotype or gender data is considered from initial ped-file. No white space allowed in phenotype entries.
No missing data allowed in gender column given in \code{sexvar}. All samples from geno file must have entries in pheno-file. 
For 1/2-coded binary phenotypes additional 0/1-coded phenotypes are generated (stored with suffix "01"). 
If gender is 1/2 (1=male, 2=female) coded in \code{sexvar}, a new variable \code{sex} 
is generated 0/1 coded (1=male, 0=female) as required by GenABEL. Optionally, numeric gonosome names can be converted to 
character names (e.g. "X", "Y"). If \code{removeNullPositions} is set TRUE, variants without valid coordinates are
removed from the generated \code{gwaa} dataset.
}
\author{
Frank Ruehle
}
