% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLP_geneRanges.R
\name{subsetByOverlaps.keepAllMeta}
\alias{subsetByOverlaps.keepAllMeta}
\title{Help functions for handling gene ranges}
\usage{
subsetByOverlaps.keepAllMeta(gr1, gr2, write.ranges.tofile = NULL,
  addStart = 0)
}
\arguments{
\item{gr1, gr2}{GRanges object}

\item{write.ranges.tofile}{character with file path. If given, a data.frame is generated
from the returned GRanges object and written to this destination. Omitted if NULL.}

\item{addStart}{numeric with mumber of bases to be added to the start coordinate
when converting a GRanges object to a dataframe. E.g. necessary for switching between 
0-based and 1-based genome coordinate systems.}
}
\value{
\code{subsetByOverlaps.keepAllMeta} returnes GRanges object containing 
overlap and meta data from input ranges. 
\code{granges2df} returns a data.frame with genomic coordinates and meta data 
from input object.
}
\description{
Keep all meta data when overlapping gene ranges or converting to dataframe.
}
\details{
By default, meta data is lost when GRanges are merged or overlapped.
\code{subsetByOverlaps.keepAllMeta} returns overlap of two GRanges objects with 
merged meta data from both. Meta data stored in \code{CompressedCharacterList}
is collapsed to a single column. 
\code{granges2df} generates a data.frame from an GRanges containing the meta data.
Meta data stored in \code{CompressedCharacterList} is collapsed to a single column.
}
\author{
Frank Ruehle
}
