% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SYB_TFsearch.R
\name{TFsearch}
\alias{TFsearch}
\title{Transcription factor binding site enrichment}
\usage{
TFsearch(sequences, newheader = NULL, annoColumn = NULL,
  name.organism = "hsapiens", projectfolder = "GEX/TFBS",
  projectname = "", applyFilter = FALSE, filtercat1 = "adj.P.Val",
  filtercat1.decreasing = FALSE, filtercat1.function = abs,
  filtercat1.threshold = 0.05, filtercat2 = "logFC",
  filtercat2.decreasing = TRUE, filtercat2.function = abs,
  filtercat2.threshold = log2(1.5), PromLookup = TRUE,
  Entrez.col = "ENTREZID", PromSeqUpstreamTSS = 2000,
  PromSeqDownstreamTSS = 200, SearchSelMotifs = NULL,
  motif.min.score = 0.9)
}
\arguments{
\item{sequences}{dataframe (or character with file path) with column containing Entrez IDs or sequence coordinates.
Required columns as defined in \code{PromLookup} (see below).}

\item{newheader}{NULL if \code{sequences} already supplied with header. Character vector with new header otherwise.}

\item{annoColumn}{(character or vector of characters) column name(s) of \code{sequences}-object with sequence annotation to maintain.}

\item{name.organism}{currently human data only (hg19).}

\item{projectfolder}{output directory.}

\item{projectname}{character prefix for output name.}

\item{applyFilter}{(boolean) If TRUE, sequences are filtered for applied categories and thresholds. Filter Values converted to ABSOLUTE values.
Optional Filtering criteria (Ignored if \code{applyFilter=FALSE}):}

\item{filtercat1}{column name of first category to filter \code{sequences} (e.g. p-values).}

\item{filtercat1.decreasing}{(boolean) direction to order \code{filtercat1}.}

\item{filtercat1.function}{select transforming function for filter category1 (no quotes). e.g. \code{abs} for absolute values, \code{identity} for no transformation}

\item{filtercat1.threshold}{Threshold for \code{filtercat1} or 'top123' for top Hits}

\item{filtercat2}{column name of second category to filter \code{sequences} (e.g. effect size).}

\item{filtercat2.decreasing}{(boolean) direction to order \code{filtercat2}.}

\item{filtercat2.function}{select transforming function for filter category2 (no quotes). E.g. \code{abs} for foldchanges}

\item{filtercat2.threshold}{Threshold for \code{filtercat2} or 'top123' for top Hits}

\item{PromLookup}{(boolean) if TRUE, all promotor sequences corresponding to genes in \code{sequences} are downloaded. Therefore a
  column with EntrezIDs is requried (column name given in \code{Entrez.col})!
if FALSE, Sequences are downloaded according to given coordinates in \code{sequences}. Therefore columns 
  for chromosome, start, stop and strand information required! Additional meta columns allowed.}

\item{Entrez.col}{name of column containing Entrez IDs}

\item{PromSeqUpstreamTSS}{definition of promotor regions to download upstream to TSS.}

\item{motif.min.score}{minimum score to match motif pwm to target sequence (ignored if SearchSelMotifs = NULL).}

\item{PromSeqDownstreamTSS:}{definition of promotor regions to download downstream to TSS.}

\item{SearchSelMotivs}{Character Vector of selected motives to search in \code{sequences} seperately to enrichment or NULL.}
}
\value{
groupReport of motifEnrichment results.
}
\description{
Function uses \code{PWMEnrich}-package for TFBS enrichment.
}
\details{
Function takes input genes, optionally looks up all promotor sequences by (unique) entrezID, 
or looks up sequences defined by coordinates, optionally filters \code{sequences} for given categories 
and calls \code{motifEnrichment} from \code{PWMEnrich}-package to identify enriched Transcription factor binding motivs. 
Optionally, tables are generated for in 'SearchSelMotifs' preselected Motivs looked up in 'sequences'.
}
\seealso{
MotifDb, PWMEnrich, PWMEnrich.Hsapiens.background
}
\author{
Frank Ruehle
}
