% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEX_wrapDESeq2.R
\name{wrapDESeq2}
\alias{wrapDESeq2}
\title{Differential expression analysis with DESeq2}
\usage{
wrapDESeq2(
  dds,
  comparisons,
  min_rowsum = 10,
  p_value_threshold = 0.05,
  adjust.method = "BH",
  fc_threshold = log2(1.5),
  projectfolder = file.path("GEX/deseq"),
  projectname = "",
  symbolColumn = NULL,
  sampleColumn = "Sample_Name",
  groupColumn = "Sample_Group",
  add_anno_columns = NULL,
  venn_comparisons = NULL,
  maxHM = 50,
  scale = c("none"),
  HMcexRow = 1.1,
  HMcexCol = 1.1,
  figure.res = 300,
  HMincludeRelevantSamplesOnly = TRUE,
  color.palette = "heat.colors"
)
}
\arguments{
\item{dds}{DESeqDataSet}

\item{comparisons}{character vector with group comparisons in format \code{"groupA-groupB"}}

\item{min_rowsum}{numeric. Minimum rowsum of countmatrix of \code{dds}. All rows with \code{rowSums < min_rowsum}
are removed from the count matrix.}

\item{p_value_threshold}{numeric p-value threshold}

\item{adjust.method}{adjustment method for multiple testing (\code{"none", "BH", "BY" and "holm"})}

\item{fc_threshold}{numeric foldchange threshold on logarithmic scale to base 2.}

\item{projectfolder}{character with directory for output files (will be generated if not exisiting).}

\item{projectname}{optional character prefix for output file names.}

\item{symbolColumn}{character with column name of Gene Symbols in \code{dds} or NULL.}

\item{sampleColumn}{character with column name of Sample names in \code{dds}}

\item{groupColumn}{character with column name of group names in \code{dds}. Group names must match \code{comparisons}!}

\item{add_anno_columns}{character with names of feature annotation columns given in \code{dds}, which shall
be added to the output tables. Omitted if \code{NULL}.}

\item{venn_comparisons}{character vector or (named) list of character vectors with group comparisons to be included 
in Venn Diagram (max 5 comparisons per Venn diagramm). Must be subset of \code{comparisons} or NULL.}

\item{maxHM}{(numeric) max number of top diff. regulated elements printed in Heatmap}

\item{scale}{character indicating if the values should be centered and scaled in either the row direction 
or the column direction, or none (default). Either \code{"none"}, \code{"row"} or \code{"column"}.}

\item{HMcexRow}{labelsize for rownames in Heatmap}

\item{HMcexCol}{labelsize for colnames in Heatmap}

\item{figure.res}{numeric resolution for png.}

\item{HMincludeRelevantSamplesOnly}{(boolean) if \code{TRUE} include only Samples in heatmap, which belong to the groups of the 
respective group comparison. If \code{FALSE}, all samples are plotted in the heatmaps. If a custom selection of sample groups 
is required for the heatmap of each group comparison, \code{HMincludeRelevantSamplesOnly} can be a named list of the form 
\code{list("groupA-groupB" = c("groupA", "groupB", "groupX"))}}

\item{color.palette}{select color palette for heatmaps}
}
\value{
List of elements 
\itemize{
  \item DEgenes list of dataframes with differential expressed genes according to significance 
  thresholds for each comparison (sorted by p-value). List elements are named by the respective group comparison. 
\item DEgenes.unfilt list of dataframes with unfiltered differential expression data for each comparison (sorted by p-value).
}
Filtered and unfiltered result tables, heatmaps, MA-plots and volcano plots for each group comparison 
as well as dispersion plot and Venn diagrams are stored as side-effects in the project folder.
}
\description{
Differential Gene expression analysis for multiple group comparisons incl. output of 
Venn diagrams, volcano plots and heatmaps.
}
\details{
Function uses the \code{DESeq2} package for differential expression analysis of an 
un-normalized, un-transformed DESeqDataSet.
Analysis is performed for any number of group comparisons. The group designations in 
the \code{groupColumn} must match the contrasts given in \code{comparisons} (\code{"groupA-groupB"}).
P-value and foldchange thresholds may be applied to filter result data. 
Volcano plots are generated for each group comparison given in \code{comparisons} with highligted 
significance thresholds.
For generation of heatmaps, the count data is log2-transformed. Heatmaps are generated for top 
differentially expressed genes for each group comparison.
}
\author{
Frank Ruehle
}
