% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MT_wrapMinfi.R
\name{wrapMinfi}
\alias{wrapMinfi}
\title{Finding differentially methylated CpG positions or regions}
\usage{
wrapMinfi(GRset, projectfolder = "MT/minfi", projectname = NULL,
  phenotype, type.covar = "categorical", returnResults = c("DMP",
  "DMR"), qCutoff = 0.001, bumpcutoff = 0.1, nResamples = 1000,
  maxGap = 500)
}
\arguments{
\item{GRset}{GenomicRatioSet or MethylSet}

\item{projectfolder}{character with directory for output files (will be generated if not exisiting).}

\item{projectname}{optional character prefix for output file names.}

\item{phenotype}{character with phenotype for differential methylation analysis. Must be given within the 
phenotype data of \code{GRset}.}

\item{type.covar}{character with type of phenotype variable, either \code{continuous} or \code{categorical}.}

\item{returnResults}{character indicating differentially methylated elements to return. 
Either \code{"DMP"} for differentially methylated positions using \code{dmpFinder}
or \code{"DMR"} for differentially methylated regions using \code{bumphunter} or both.}

\item{qCutoff}{numeric with q-value (FDR) threshold for DMPs to be reported by \code{dmpFinder}.}

\item{bumpcutoff}{numeric start value to find the appropriate \code{bumphunter} cutoff.
A cutoff 0.1 corresponds to 10% difference on the Beta-values.
If the cutoff results in more than 10000 candidate DMRs, the cutoff is automatically increased
before applying permutation testing to avoid excessive computation time.}

\item{nResamples}{numeric, number of resamples to use when computing null distributions with \code{bumphunter}.}

\item{maxGap}{numeric with maximum location gap used to define clusters of probes for \code{bumphunter}.}
}
\value{
list with up to 2 elements depending on selection in \code{returnResults}. 
        Result tables are also stored as side-effects in \code{projectfolder}
\itemize{
  \item \code{DMP} dataframe of CpG positions sorted by differential methylation p-value annotated by respective annotation package.
  \item \code{DMR} An object of class \code{bumps} with the components: \code{tab, coef, fitted, pvaluesMarginal, null, algorithm}.
                   See help(bumphunter) for details. The dataframe in \code{tab} is annotated with meta data for all CpG-sites residing
                   within the respective DMR.
}
}
\description{
Use \code{minfi}-package to determine either differentially methylated positions (DMPs, \code{dmpFinder}) and/or 
differentially methylated regions (DMRs, \code{bumphunter}).
}
\details{
The \code{dmpFinder} function from \code{minfi} identifies differentially methylated CpG-sites using
linear regression for continuous phenotypes and F-test for categorical phenotypes. In case of a categorical
phenotype with many groups (e.g. experimental groups), F-test is applied over all groups. 
Tests are performed on M values, which are logit transformed Beta values 
(beta = Methylated allele intensity / (Unmethylated allele intensity + Methylated allele intensity + 100)).
Annotation data of \code{GRset} is added to the result table.

Instead of looking for association between a single genomic location and a phenotype of interest, 
\code{bumphunter} looks for genomic regions that are differentially methylated (beta values). In the context of the 450k array, 
the algorithm first defines clusters of probes such that two consecutive probe locations in the cluster
are not separated by more than distance \code{mapGap}. Briefly, the algorithm first computes a t-statistic for beta values
at each genomic location, with optional smoothing. Then, it defines a candidate region to be a cluster of probes
for which all the t-statistics exceed a predefined threshold. To test for significance of the candidate regions, 
the algorithm uses permutations given in \code{nResamples}.
}
\author{
Frank Ruehle
}
