% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SYB_wrapPWMEnrich.R
\name{wrapPWMEnrich}
\alias{wrapPWMEnrich}
\title{Transcription factor binding site enrichment}
\usage{
wrapPWMEnrich(
  sequences,
  newheader = NULL,
  annoColumn = NULL,
  name.organism = "hsapiens",
  projectfolder = "GEX/TFBS",
  projectname = "",
  figure.res = 300,
  applyFilter = FALSE,
  filtercat1 = "adj.P.Val",
  filtercat1.decreasing = FALSE,
  filtercat1.function = abs,
  filtercat1.threshold = 0.05,
  filtercat2 = "logFC",
  filtercat2.decreasing = TRUE,
  filtercat2.function = abs,
  filtercat2.threshold = log2(1.5),
  PromLookup = TRUE,
  id.type = "ENTREZID",
  id.column = "ENTREZID",
  PromSeqUpstreamTSS = 2000,
  PromSeqDownstreamTSS = 200,
  SearchSelMotifs = NULL,
  motif.min.score = 0.9
)
}
\arguments{
\item{sequences}{dataframe or character with file path to dataframe or named list containing dataframes. The dataframe must either contain
a column with Entrez IDs or sequence coordinates if \code{PromLookup == FALSE}. Required columns for latter case are defined 
in \code{PromLookup} below.}

\item{newheader}{NULL if \code{sequences} already supplied with header. Character vector with new header otherwise.}

\item{annoColumn}{character or vector of characters. Column name(s) of \code{sequences}-object with sequence annotation to maintain.}

\item{name.organism}{currently human data only (hg19).}

\item{projectfolder}{output directory.}

\item{projectname}{character prefix for output name.}

\item{figure.res}{numeric resolution for png.}

\item{applyFilter}{(boolean) If TRUE, sequences are filtered for applied categories and thresholds. Filter Values converted to ABSOLUTE values.
Optional Filtering criteria (Ignored if \code{applyFilter=FALSE}):}

\item{filtercat1}{column name of first category to filter \code{sequences} (e.g. p-values).}

\item{filtercat1.decreasing}{(boolean) direction to order \code{filtercat1}.}

\item{filtercat1.function}{select transforming function for filter category1 (no quotes). e.g. \code{abs} for absolute values, \code{identity} for no transformation}

\item{filtercat1.threshold}{Threshold for \code{filtercat1} or 'top123' for top Hits}

\item{filtercat2}{column name of second category to filter \code{sequences} (e.g. effect size).}

\item{filtercat2.decreasing}{(boolean) direction to order \code{filtercat2}.}

\item{filtercat2.function}{select transforming function for filter category2 (no quotes). E.g. \code{abs} for foldchanges}

\item{filtercat2.threshold}{Threshold for \code{filtercat2} or 'top123' for top Hits}

\item{PromLookup}{(boolean) if TRUE, all promotor sequences corresponding to genes in \code{sequences} are downloaded. 
Therefore a column with EntrezIDs is requried (column name given in \code{Entrez.col}). if FALSE, Sequences are 
downloaded according to given coordinates in \code{sequences}. Therefore columns for \code{chromosome}, \code{start}, 
\code{stop} and \code{strand} information are required! Additional meta columns allowed.}

\item{id.type}{character with identifier type from annotation package ("ENTREZID" or "SYMBOL")
Gene symbols Will be converted to EntrezIDs prior to enrichment analysis.}

\item{id.column}{character with column name for identifier variable in \code{sequences}.}

\item{PromSeqUpstreamTSS}{definition of promotor regions to download upstream to TSS.}

\item{PromSeqDownstreamTSS}{definition of promotor regions to download downstream to TSS.}

\item{SearchSelMotifs}{Character Vector of selected motives to search in \code{sequences}. Omitted if NULL.}

\item{motif.min.score}{minimum score to match motif pwm to target sequence (ignored if SearchSelMotifs = NULL).}
}
\value{
groupReport of motifEnrichment results.
}
\description{
This function uses the \code{PWMEnrich}-package for TFBS enrichment.
}
\details{
The function takes input genes in \code{sequences} and looks up all promotor sequences by (unique) entrezIDs refering to human genome build hg19.
If \code{PromLookup == FALSE}, look up of promotor sequences is omitted and sequences of interest must be given in \code{sequences}
as coordinates instead. The dataset may be filtered for designated filter criteria if desired. 
These sequences are transferred to \code{motifEnrichment} from \code{PWMEnrich}-package to identify enriched Transcription factor 
binding motivs for the input sequences. Optionally, preselected motivs given in \code{SearchSelMotifs} are looked up in \code{sequences}. 
All result tables and plots are stored in the project folder.
}
\seealso{
MotifDb, PWMEnrich, PWMEnrich.Hsapiens.background
}
\author{
Frank Ruehle
}
