% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SYB_wrapRDAVID.R
\name{wrapRDAVID}
\alias{wrapRDAVID}
\title{Gene enrichment analysis using the DAVID Database}
\usage{
wrapRDAVID(
  genes,
  newheader = NULL,
  backgroundlist = NULL,
  newheaderBackground = NULL,
  davidAccount = NULL,
  projectfolder = "GEX/DAVID_Enrichment",
  projectname = "",
  maxDavid = 100,
  davidCat = c("GOTERM_BP_ALL", "GOTERM_MF_ALL", "GOTERM_CC_ALL", "KEGG_PATHWAY",
    "BIOCARTA", "UCSC_TFBS"),
  id.type = "ENTREZ_GENE_ID",
  id.column = "ENTREZID",
  sortcolumn = "adj.P.Val",
  sortdecreasing = FALSE,
  sortcolumn.threshold = 0.05,
  org = "Homo sapiens",
  threshold4enrich = 0.05,
  mincount = 2L
)
}
\arguments{
\item{genes}{vector with gene names to analyse or dataframe or character with path to dataframe. Supplied
dataframe needs to include a column with identifiers specified in \code{id.column}.}

\item{newheader}{optional character vector with new header information for \code{genes} dataframe. Only relevant 
if \code{genes} is a dataframe or character string with filepath to a table with wrong or missing header. 
NULL otherwise.}

\item{backgroundlist}{optional dataframe containing background list for enrichment analysis. 
If NULL, species specific background from DAVID is used instead.}

\item{newheaderBackground}{optional character vector with new header information for \code{backgroundlist} dataframe.}

\item{davidAccount}{email account for DAVID's Web Service connectivity or DAVIDWebService object.}

\item{projectfolder}{character with directory for output files (will be generated if not exisiting).}

\item{projectname}{optional character prefix for output file names.}

\item{maxDavid}{(numeric) max number of top diff regulated elements used for enrichment analysis.}

\item{davidCat}{character vector with DAVID categories to enrich for.}

\item{id.type}{character with identifier DAVID uses to enrich for (e.g. "ENTREZ_GENE_ID" or "OFFICIAL_GENE_SYMBOL")}

\item{id.column}{character with respective column name for \code{id.type} in \code{genes}.}

\item{sortcolumn}{character with column name to be sorted for selecting top entries.}

\item{sortdecreasing}{(boolean) drecreasing parameter for order of \code{sortcolumn}.}

\item{sortcolumn.threshold}{numeric threshold for \code{sortcolumn} to be included in overepresentation analysis.
\code{If sortdecreasing=F, value < sortcolumn.threshold 
else value > sortcolumn.threshold}}

\item{org}{character with species name DAVID uses for enrichment (e.g."Homo sapiens" or "Mus musculus").}

\item{threshold4enrich}{numeric p-value threshold for enrichment terms.}

\item{mincount}{numeric minimum gene count in enrichment terms.}
}
\value{
Functional annotation chart as R object. 
Additionally, a functional annotation chart is stored in the project folder as side effect.
}
\description{
Wrapper for gene enrichment analysis using the Database for Annotation, Visualization and Integrated Discovery (DAVID) 
accessed by the \code{RDAVIDWebService}
}
\details{
Function uses genelist and optionally background gene list as input to perform enrichment analysis 
using the RDAVIDWebService. Registration for RDAVIDWebService is required at 
\code{https://david.ncifcrf.gov/webservice/register.htm}.
Gene list can be supplied as character vector with ids given in \code{id.type}, as dataframe
or as character string with path of data file. If a vector is given, sorting parameter will be ignored.
Enrichment categories for a defined organism are defined in \code{davidCat}.
If no connection to \code{RDAVIDWebService} can be established, NULL is returned.
}
\references{
https://david.ncifcrf.gov/
}
\author{
Frank Ruehle
}
