% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLP_basic_gene_annotation.R
\name{basicAnno}
\alias{basicAnno}
\title{Basic gene annotation}
\usage{
basicAnno(data, Symbol.column = "SYMBOL", Entrez.column = "ENTREZID",
  org = "human")
}
\arguments{
\item{data}{dataframe or character with filepath to data to be loaded.}

\item{Symbol.column}{character with column name of Gene Symbols in \code{data}.
if NULL, SYMBOLs are derived from \code{Entrez.column} using the annotation package for \code{org}.}

\item{Entrez.column}{character with column name of ENTREZ IDs in \code{data}.
if NULL, ENTREZ IDs are derived from \code{Symbol.column} using the annotation package for \code{org}.
Either \code{Symbol.column} or \code{Entrez.column} must be specified.}

\item{org}{character with species name ("human", "mouse", "rat").}
}
\value{
dataframe with added annotation column.
}
\description{
Annotate gene symbols and EntrezIDs using an organism-specific annotation package.
}
\details{
This function adds annotation of either gene symbol or EntrezID to \code{data}
using the corresponding annotation package of the organism specified in \code{org}.
}
\author{
Frank Ruehle
}
