% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLP_makeDomainsFromExons.R
\name{makeDomainsFromExons}
\alias{makeDomainsFromExons}
\title{Genomic coordinates of protein domains}
\usage{
makeDomainsFromExons(ID, biomaRt, uniprot_domains.gff,
  suffix.outputFilname = ".txt")
}
\arguments{
\item{ID}{cahracter with either Ensembl transcript ID, Ensembl gene ID or gene symbol.}

\item{biomaRt}{biomaRt object to obtain exon ccordinates.}

\item{uniprot_domains.gff}{character with file path to Uniprot gff export. 
In Uniprot, select desired features e.g. \code{PTM/Processing} and \code{Family & Domains} 
and export the basket to gff-file format.}

\item{suffix.outputFilname}{character to be added at output filename.}
}
\value{
dataframe with exon data downloaded from \code{biomaRt}. 
The result domain table is stored as side effect in the filepath given by \code{uniprot_domains.gff}.
}
\description{
Use genomic exon coordinates and Uniprot protein domain data to determine genomic 
coordinates of protein domains for plotting in genomic context.
}
\details{
Genomic exon coordinates for the desired transcript are doenloaded from biomaRt. If a gene is selected,
the canonical transcript is determined as transcript with longest coding sequence.
Protein data from Uniprot is expected in gff format. Any kind of domains or features may be selected.
The strored result table may be edited later to modify suggested plotting parameter.
}
\author{
Frank Ruehle
}
