% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEX_readData2eset.R
\name{readData2eset}
\alias{readData2eset}
\title{Create ExpressionSet object from expression data}
\usage{
readData2eset(exprsData, phenoData, featureData = NULL,
  ProbeID = "PROBE_ID", sampleColumn = "Sample_Name",
  groupColumn = "Sample_Group", experimentData = MIAME(),
  arrayAnnotation = "Humanv4", transform = log2, organism = "human")
}
\arguments{
\item{exprsData}{dataframe or character with filepath to expression data to be loaded.
If no rownames are included, a column containing ProbeIDs is expected and used as feature names.}

\item{phenoData}{dataframe or character with filepath to sample sheet / phenotype data to be loaded.}

\item{featureData}{dataframe or character with filepath to feature data to be loaded.
If NULL, non-numeric data columns from 'exprsData' are used as feature data (if any).}

\item{ProbeID}{character with name of the column in 'exprsData' and 'featureData' that contains 
identifiers used to uniquely identify each probe.}

\item{sampleColumn}{character with name of sample column in 'phenoData'.}

\item{groupColumn}{character with name of group column in 'phenoData'.}

\item{experimentData}{MIAME object with optional experiment description.}

\item{arrayAnnotation}{character string specifying array annotation package (e.g. "Humanv4"). character() if not available.}

\item{transform}{optional function definition to transform expression data (e.g. log2 for logarithm of base 2).
if NULL, no data transformation is performed.}
}
\value{
ExpressionSet object
}
\description{
ExpressionSet object assembled from expression, feature and phenotype data.
}
\details{
Expression data, phenotype data and optional feature data are used to make an ExpressionSet object.
Only numeric columns in \code{exprsData} are used. If no feature data is given, non-numeric data columns 
from \code{exprsData} are used as feature data (if any). If dataframes don't contain rownames as identifier, 
the column \code{ProbeID} is used as rownames in \code{exprsData} and \code{featureData} and column \code{sampleColumn} in \code{phenoData}.
Rownames of \code{exprsData} must match with rownames of \code{featureData} (if given). Column names of \code{exprsData}
must match with rownames of \code{phenoData}. Characters "/" and "-" in sample names or group names are 
replaced by ".". If either gene symbols or EntrezIDs are missing in feature data, this annotation is
added using the corresponding annotation package (species name is derived from \code{arrayAnnotation}). The optional
Function given in \code{transform} is used to transform the expression data.
}
\author{
Frank Ruehle
}
