% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SYB_wrapClusterProfiler.R
\name{wrapClusterProfiler}
\alias{wrapClusterProfiler}
\title{Wrapper for gene enrichment analysis using clusterProfiler}
\usage{
wrapClusterProfiler(genes, newheader = NULL, backgroundlist = NULL,
  newheaderBackground = NULL, projectfolder = "GEX/clusterProfiler",
  projectname = "", enrichmentCat = c("GO", "KEGG", "Reactome", "DO"),
  maxInputGenes = 100, id.type = "ENTREZID", id.column = "ENTREZID",
  sortcolumn = "adj.P.Val", sortdecreasing = FALSE,
  sortcolumn.threshold = 0.05, fun.transf.incr.values = function(x) {    
  -log10(x) }, FCcolumn = "logFC", threshold_FC = log2(1.5),
  org = "human", pAdjustMethod = "BH", enrich.p.valueCutoff = 0.05,
  enrich.q.valueCutoff = 0.05, nPerm = 1000, minGSSize = 10,
  maxGSSize = 500, figure.res = 300)
}
\arguments{
\item{genes}{vector with gene names to analyse or dataframe or character with path to dataframe or list thereof. 
Supplied dataframe needs to include a column with identifiers specified in \code{id.column}.}

\item{newheader}{optional character vector with new header information for \code{genes} dataframe. Only relevant 
if 'genes' is a dataframe (or character string with filepath to a table) with wrong or missing header. 
NULL otherwise.}

\item{backgroundlist}{optional background IDs for enrichment analysis. Can either be a vector with gene IDs 
or dataframe or character with path to dataframe. If given, an \code{id.column} is needed as 
in \code{genes} with type of IDs given in \code{id.type}. 
If "genome", all ENTREZ IDs from the annotation package of the respective organism
(denoted in \code{org}) are used as background.
If NULL, full ID list from \code{genes} is used as background (i.e. that \code{genes} need to 
contain all genes under investigation without pre-filtering for significance).}

\item{newheaderBackground}{optional character vector with new header information for \code{backgroundlist}.}

\item{projectfolder}{character with directory for output files (will be generated if not existing).}

\item{projectname}{optional character prefix for output file names.}

\item{enrichmentCat}{character vector with categories to be enriched (\code{GO}: gene ontology (MF, BP, CC), 
\code{KEGG}: KEGG pathways, \code{Reactome}: Reactome pathways, \code{DO}: Disease ontology).
Disease ontology is for human only.}

\item{maxInputGenes}{(numeric) max number of top diff regulated elements used for enrichment analysis (or NULL).}

\item{id.type}{character with identifier type from annotation package (\code{"ENTREZID"} or \code{"SYMBOL"})
Gene symbols Will be converted to EntrezIDs prior to enrichment analysis.}

\item{id.column}{character with column name for identifier variable in \code{genes}.}

\item{sortcolumn}{character with column name of quantitative data in \code{genes} used for ordering.
If Null, ranking of genes is omitted and GSEA not possible.}

\item{sortdecreasing}{(boolean) order parameter for hierarchy of values in \code{sortcolumn}.
FALSE for increasing values (e.g. p-values), TRUE for decreasing values (e.g. fold changes).
If FALSE, values in sortcolumn will be transformed prior to GSEA.}

\item{sortcolumn.threshold}{numeric threshold for \code{sortcolumn} to be included in over-representation analysis.
\code{If sortdecreasing=F, value < sortcolumn.threshold 
else value > sortcolumn.threshold}}

\item{fun.transf.incr.values}{function definition for transforming values in \code{sortcolumn} prior to GSEA.
Is applied if \code{sortdecreasing=F}.}

\item{FCcolumn}{(character) optional column name of foldchanges in \code{genes} if \code{sortcolumn} is used elsewhere.
Used only for cnetplot of enrichment results. Omitted if NULL}

\item{threshold_FC}{(numeric) Fold change threshold for filtering (threshold interpreted for log transformed foldchange values!)
Only relevant for over-representation analysis if an unfiltered gene list is given in \code{genes}
to allow for parallel GSEA.}

\item{org}{character with name of organism ("human", "mouse", "rat").}

\item{pAdjustMethod}{method for adjusting for multiple testing. 
One of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"}

\item{enrich.p.valueCutoff}{numeric p-value threshold for returned enrichment terms.}

\item{enrich.q.valueCutoff}{numeric q-value threshold for returned enrichment terms.}

\item{nPerm}{permutation numbers for gene set enrichment analysis}

\item{minGSSize}{minimal size of genes annotated by Ontology term for testing.}

\item{maxGSSize}{maximal size of genes annotated for testing}

\item{figure.res}{numeric resolution for png.}
}
\value{
List of enrichment-objects defined in \code{DOSE}-package (\code{enrichResult}-object for overrepresentation 
  analysis, \code{gseaResult}-objects for gene set enrichment analysis).                          
  Enrichment tables and plots are stored in the project folder as side effects.
}
\description{
Applying over-representation analysis and gene set enrichment analysis to supplied gene list
}
\details{
Function uses genelist and optionally background gene list as input to perform enrichment analysis 
using the \code{clusterProfiler}-package. By default, over-representation analysis and gene set enrichment analysis 
(GSEA) is performed for all categories given in \code{enrichmentCat}.
The Gene list can be supplied as vector with ids given in \code{id.type},
dataframe with IDs of type \code{id.type} given in column \code{id.column} or character string with path of data file.
The IDs are converted to ENTREZ IDs (if necessary) prior to enrichment using the annotation package
for the species denoted in \code{org}.
Optionally, quantitative data can be included in \code{sortcolumn} for sorting and filtering (using \code{sortcolumn.threshold}) 
the data. If quantitative data are no fold changes, fold changes may be given additionally in \code{FCcolumn}.
This data is used for cnetplot of enrichment results as well as for KEGG pathway mapping of enriched 
KEGG pathways (if any) using the \code{pathview} package.
If \code{FCcolumn == NULL} the data in \code{sortcolumn} is also used for the cnetplots and pathway mapping. 
Be aware that the legend of the cnetplots will be "Fold Change" anyway.
If no quantitative data is provided, sorting and GSEA will be skipped.
Optionally, a background list can be provided for enrichment analysis.
}
\author{
Frank Ruehle
}
